Week 2: Assignment 2 - Question 1
You are given a non-negative sequence of numbers, ending with a -1. You can
assume that there are at least two numbers before the ending -1. 
You have to output the second largest element of the sequence.
If there is no second largest element in the sequence then output 0.
Note : -1 is not a part of input. It only signifies that input has ended.