Week 2: Assignment 2 - Question 3
In this assignment, you will be given an  N×N
 matrix, with N>1

You have to determine whether the matrix is an upper triangular matrix.

A matrix is upper triangular if every entry below the diagonal is 0. The following is an example of an upper triangular matrix:
1000
1400
1100
1111
Note: The diagonal itself, and the entries above the diagonal can be zeroes or non-zero integers.