Week 5: Assignment 5 - Question 1
Write a recursive program that inputs a line of characters from the user.
The line may contain blanks. It outputs the line with the characters reversed. The input ends with EOF (end of file).

Hint: Use getchar() to read input and check for c != EOF. 

NOTE: You have to use recursion to solve this, and are NOT allowed to use array to store the input!!