Week 5: Assignment 5 - Question 3
Write a recursive code to find the block product of an array.

The following is the algorithm to find the block product recursively:

Given an array A partition it into four quadrants of equal size:

                          M=[ABCD]


The block product of the array is defined recursively as : 

                           |M|=|A|∗|D|−|B|∗|C|


With the base case, when n = 4 ,  |[a,b,c,d]|=ad−bc


Note : You can assume that n is a power of 4. n>=4 and n <= 1024.