Week 6: Assignment 6 - Question 1
We say that a string 's' is an anagram of another string 't' if the letters in 's' can be rearranged to form 't'.

For example, "butterfly" is an anagram of "flutterby", since a rearrangement of the first word results in the second.

We say that a position 'i' in 's' and 't' match, if 's' is an anagram of 't', and s[i]==t[i].

In this question, you will be given two words, 's' and 't'. You have to output the number of matching positions if s is an anagram of t, and -1 if s is not an anagram of t.