Week 7: Assignment 7 - Question 2
You have to complete the code for performing the following operations in a priority queue linked list:
Create and return a node e with given id and value val
struct node * create_node(int id, int val);

Add an node e to the beginning of the list. Return the new list.
struct node * append(struct node * list, struct node * e);

Search for a node e with id inside the list. Return a pointer to e  if found, else return NULL
struct node * search(struct node * list, int id); 

Change the value of an element e, (if found) in the list to the new value val. 
void change_priority((struct node *list, struct node* e, int val);
Each element in the linked list is a structure of the following format:

struct node{
	int id;
	int priority;
	struct node *next;
}; 

The field priority denotes the priority of an element inside the list, the higher the value of the field, the more the priority.
Element with the highest priority is returned by extract_max.

Note: The code for manipulating the input as well as output is given to you. You only have to write code for the incomplete functions.